Copy over changes from `EntityTag` branch
